Abstract out user and team service
Moves the user and team services into the shared module for better lazy
loading support. This also creates the user and team interfaces and
moves all other classes to use the new interfaces the new services.

Using the abstracted user service, this adds the ability for the create
a team component to populate the team select box with all of the users
that are stored in the db rather than entering text.

This also adds an initial version of a teams list page.